Set up some basic IAM using Terraform.
Creating 4 users with terraform, a group, policy (and attaching the policy). Accounts need to generate a new password manually. They are added to the group automatically, which has administrator access, but it is not possible to log into the accounts.

As of 4/2017 it isn't reccomended to store tfstate files in version control per here :

https://github.com/hashicorp/terraform/issues/13891

It is also reccomended not to store the .terraform folder into version control per here :

https://www.terraform.io/docs/commands/get.html

These should be put in the S3 bucket instead.