Getting certs, ALB ready.
The instances are taking longer than planned to get running, so I have started working on the ALB listener for HTTP and HTTPS and target groups. These should be basically ready to go once the instances are running. Certs may take a little more work because they require validation thorugh our DNS or through an email.